requirements.txt
streamlit 
streamlit_lottie
pymongo 
plotly 
networkx 
pyvis 
requests 
python-dotenv